factored out userbadge code from main flask app
