+++ lib/UserBadge/__init__.py
@@ -1,0 +1,3 @@
+__version__ = '0.1'
+
+from _user_badge import UserBadge, RankCodeAbbey
